Refactor AI package layout into pipeline subdirectories
Move utility, social, and tactical AI modules into staged src subdirectories and rewire imports/exports without changing behavior, establishing a reviewable first slice for issue #2289.
