+++ packages/colonizethis_ai/lib/colonizethis_ai.dart
@@ -16,9 +16,9 @@ export 'src/diplomacy_planner.dart';
 export 'src/economy_planner.dart';
 export 'src/full_ai_planner.dart';
 export 'src/goal_manager.dart';
-export 'src/hidden_agenda.dart';
-export 'src/ai_validation_exception.dart';
-export 'src/mood_state_machine.dart';
+export 'src/social/hidden_agenda.dart';
+export 'src/util/ai_validation_exception.dart';
+export 'src/social/mood_state_machine.dart';
 export 'src/perception.dart';
 export 'src/strategic_ai.dart';
-export 'src/tactical_ai.dart';
+export 'src/tactical/tactical_ai.dart';
