+++ packages/colonizethis_ai/lib/src/diplomacy_planner.dart
@@ -6,10 +6,10 @@ import 'package:colonizethis_logic/ai_api.dart';
 import 'package:colonizethis_logic/order_suggestion_api.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import 'ai_random_utils.dart';
+import 'util/ai_random_utils.dart';
 import 'goal_manager.dart';
-import 'hidden_agenda.dart';
-import 'orders_extensions.dart';
+import 'social/hidden_agenda.dart';
+import 'util/orders_extensions.dart';
 import 'perception.dart';
 
 final _log = packageLogger();
