+++ packages/colonizethis_ai/lib/src/domain_planners.dart
@@ -6,11 +6,11 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_logic/ai_api.dart';
 import 'package:colonizethis_logic/order_suggestion_api.dart';
 
-import 'ai_random_utils.dart';
+import 'util/ai_random_utils.dart';
 import 'diplomacy_planner.dart';
 import 'goal_manager.dart';
-import 'hidden_agenda.dart';
-import 'orders_extensions.dart';
+import 'social/hidden_agenda.dart';
+import 'util/orders_extensions.dart';
 import 'perception.dart';
 
 final _log = packageLogger();
