+++ packages/colonizethis_ai/lib/src/social/mood_state_machine.dart
