+++ packages/colonizethis_ai/lib/src/util/ai_random_utils.dart
