+++ packages/colonizethis_ai/lib/src/util/ai_validation_exception.dart
