+++ packages/colonizethis_ai/lib/src/util/orders_extensions.dart
@@ -1,4 +1,3 @@
-import 'package:colonizethis_logic/order_suggestion_api.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
 typedef _OrdersByPlayer<T> = Map<String, List<T>>;
